+++ test/unit/specs/util/test-schemas.js
@@ -809,3 +809,24 @@ export const createRowSchema = {
     }
   ]
 }
+
+export const showNonVisibleAttributeSchema = {
+  'attributes': [
+    {
+      'attributes': [],
+      'auto': false,
+      'fieldType': 'STRING',
+      'href': '/api/v2/sys_sec_User/meta/username',
+      'isAggregatable': false,
+      'label': 'Username',
+      'labelAttribute': true,
+      'lookupAttribute': true,
+      'maxLength': 255,
+      'name': 'username',
+      'nillable': true,
+      'readOnly': false,
+      'unique': false,
+      'visible': false
+    }
+  ]
+}