+++ CHANGELOG.md
@@ -5,6 +5,10 @@
     * Remove 'number' type ( ***breaking change***)
     * Add support for mapping int, long and decimal to EntityMapper
 * Update dev/demo/e2e-test setup, add page per feature to improve development, debugging and e2e-test experience 
+* Add showNonVisibleAttributes option to EntityMapper options 
+    * If set to true maps non visible attribute to visible field
+    * showNonVisibleAttributes is optional and defaults to false
+    * If visible expression is set this takes precedence 
        
 
 <a name="0.12.0"></a>